// _interface/types.ts
export type TStatus = 'add'
  | 'remove';
export type TLanguage = 'en'
  | 'ru'
  | 'uk'
  | 'fr';
export type TPageId = 'nav'
  | 'contacts'
  | 'about'
  | 'diploma'
  | 'service'
  | 'article'
  | 'main'
  | 'review'
  | '[modal] appointment'
  | '[modal] review'
  | 'index'
  | 'article-item';
export type TWeekday = 'su'
  | 'mo'
  | 'tu'
  | 'we'
  | 'th'
  | 'fr'
  | 'sa';